+++ packages/f-header/src/tenants/es-ES.js
@@ -48,12 +48,12 @@ export default {
     },
     accountLogin: {
         text: 'Inicia sesión',
-        url: '/account/login?returnurl=',
+        url: '/account/login',
         gtm: 'click_login'
     },
     accountLogout: {
         text: 'Salir',
-        url: '/account/logout?returnurl=',
+        url: '/account/logout',
         gtm: 'click_logout'
     }
 };